requirements.txt
annotated-types==0.7.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
idna==3.10
markdown-it-py==4.0.0
mdurl==0.1.2
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
requests==2.32.5
rich==14.1.0
shellingham==1.5.4
SQLAlchemy==2.0.43
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
openai
anthropic
google-generativeai
mistralai
cohere
fastapi
uvicorn[standard]
websockets
psycopg[binary]>=3.1
alembic>=1.13